pyproject.toml
[project]
name = "rag-agent"
version = "0.1.0"
description = "RAG Agent System for Legal Question Answering"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "pydantic>=2.9.0",
    "openai>=1.54.0",
    "pandas>=2.2.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "chromadb>=0.5.0",
    "tqdm>=4.66.0",
    "rank-bm25>=0.2.2",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.8.0",
    "pyright>=1.1.389",
    "pytest>=8.3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

